requirements.txt
black>=25.1.0
build>=1.2.1
flake8>=7.3.0
jsonargparse[signatures]>=4.40.0,<5.0.0
lightning==2.5.2
numpy<=2.3.1
pytest>=8.4.2
PyYAML>=6.0.2,<7.0.0
scipy>=1.15.3
torch==2.7.1
torchmetrics==1.7.4
twine>=6.1.0
wandb>=0.18.7,<0.19.0
yoyodyne==0.4.4
